refactor: insert files into lib folder
