+++ src/lib/__mocks__/block.ts
@@ -1,17 +1,17 @@
-import BlockInfo from "../blockinfo";
-import Validation from "../lib/validation";
+import Validation from "../validation";
+import Transaction from "./transaction";
 
 export default class Block {
   index: number;
   hash: string;
   timestamp: number;
   previousHash: string;
-  data: string;
+  transactions: Transaction[];
 
   constructor(block?: Block) {
     this.index = block?.index || 0;
     this.previousHash = block?.previousHash || "";
-    this.data = block?.data || "";
+    this.transactions = block?.transactions || [];
     this.timestamp = block?.timestamp || Date.now();
     this.hash = block?.hash || this.getHash();
   }
@@ -24,13 +24,5 @@ export default class Block {
     if (!previusHash || previusIndex < 0 || this.index < 0)
       return new Validation(false, "Invalid mock block.");
     return new Validation();
-  }
-
-  static fromBlockInfo(blockInfo: BlockInfo): Block {
-    const block = new Block();
-    block.index = blockInfo.index;
-    block.previousHash = blockInfo.previousHash;
-    block.data = blockInfo.data;
-    return block;
   }
 }
