+++ src/lib/blockchain.ts
@@ -1,7 +1,7 @@
 import Block from "./block";
 import BlockInfo from "./blockinfo";
-import Transaction, { TransactionType } from "./lib/transaction";
-import Validation from "./lib/validation";
+import Transaction, { TransactionType } from "./transaction";
+import Validation from "./validation";
 
 export default class BlockChain {
   blocks: Block[];
